LayoutTests/media/wireless-playback-media-player/volume-expected.txt


Test that setting video.volume propagates to a MediaDeviceRoute.

RUN(internals.mockMediaDeviceRouteController.activateRoute(route))
RUN(video.src = findMediaFile('video', '../content/test'))
RUN(video.play())
RUN(route.ready = true)
EVENT(canplaythrough)
RUN(internals.setPageMediaVolume(1))
RUN(internals.setMediaElementVolumeLocked(video, false))
RUN(video.volume = 0.5)
EXPECTED (route.volume == '0.5') OK
END OF TEST
